feat: move kafka behind a build tag so the default build is cgo-free
