+++ internal/infrastructure/codec/avro_test.go
@@ -1,3 +1,5 @@
+//go:build kafka
+
 package codec
 
 import (
