+++ internal/infrastructure/kafka/consumer.go
@@ -1,3 +1,5 @@
+//go:build kafka
+
 package kafka
 
 import (
