+++ internal/infrastructure/kafka/producer.go
@@ -1,3 +1,5 @@
+//go:build kafka
+
 package kafka
 
 import (
