+++ internal/infrastructure/kafka/sasl.go
@@ -1,19 +1,10 @@
 package kafka
 
-import (
-	ckf "github.com/confluentinc/confluent-kafka-go/v2/kafka"
-)
-
+// SASLConfig carries Kafka SASL/PLAIN credentials. The struct is always
+// compiled (referenced by eventbus.KafkaConfig and the service mains) even in
+// the default CGO-free build; the confluent-specific Generate() method lives in
+// sasl_kafka.go behind the `kafka` build tag.
 type SASLConfig struct {
 	Username string
 	Password string
-}
-
-func (c *SASLConfig) Generate() map[string]ckf.ConfigValue {
-	return map[string]ckf.ConfigValue{
-		"sasl.mechanisms":   "PLAIN",
-		"security.protocol": "SASL_SSL",
-		"sasl.username":     c.Username,
-		"sasl.password":     c.Password,
-	}
 }
